refactored backend files into Server folder
