--- buzzmoon-frontend/buzzmoon-vite/src/Components/Home/Home.jsx
@@ -1,12 +1,0 @@
-import * as React from 'react';
-import CompeteView from '../CompeteView/CompeteView';
-import Navbar from '../Navbar/Navbar';
-
-export default function Home(props) {
-  return (
-    <div className='home'>
-      <Navbar setCurrentUser={props.setCurrentUser}/>
-      <CompeteView />
-    </div>
-  );
-}
